Screenshots now appear as a grid when multiple are taken
